pyproject.toml
[tool.poetry]
name = "send-data-to-postgresql"
version = "0.1.0"
description = ""
authors = ["levyvix <levy.vix@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
faker = "^26.0.0"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
faker-commerce = "^1.0.4"
schedule = "^1.2.2"
kafka-python = "^2.0.2"
confluent-kafka = "^2.4.0"
fastavro = "^1.9.4"
avro-python3 = "^1.10.2"
py-ccloud = "^0.1.2"
requests = "^2.32.3"
rich = "^13.7.1"
pyspark = "^3.5.1"
duckdb = "^1.0.0"
pandas = "^2.2.2"
deltalake = "^0.18.2"
pyarrow = "^16.1.0"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
